fix cloudflare core module cycle
