+++ cloudflare/packages/core/src/service.ts
@@ -11,10 +11,7 @@
  * all. The subrouter just routes; it does not gate.
  */
 import { Context, Effect, Layer } from "effect"
-import {
-  AccountStoreTag,
-  type Account,
-} from "./index.ts"
+import { AccountStoreTag, type Account } from "./account-store.ts"
 
 export class NoEligibleAccount extends Error {
   readonly _tag = "NoEligibleAccount"
